checkout

+++ src/main/java/com/example/demospringncc/dto/ProductDTO.java
@@ -8,8 +8,8 @@ public class ProductDTO {
     private  Long id;
     private  String name;
     private int categoryId;
-    private double price;
-    private  double weight;
+    private long price;
+    private  long weight;
     private String description;
     private String imageName;
 }
